pyproject.toml
[project]
name = "star-games"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

description = "Add your description here"
dependencies = [
  "aiofiles>=24.1.0",
  "click>=8.1.8",
  "easydict>=1.13",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "jsonrpcserver>=5.0.9",
  "platformdirs>=4.3.6",
  "pywebview>=5.3.2",
  "uvicorn>=0.34.0",
]


[project.optional-dependencies]
gtk = ["pywebview[gtk]>=5.3.2"]
qt = ["pywebview[qt]>=5.3.2"]

[dependency-groups]
dev = ["pyinstaller>=6.14.0"]
